+++ com.sparta.logistics.auth/src/main/java/com/sparta/logistics/auth/libs/jwt/JwtUtil.java
@@ -1,4 +1,4 @@
-package com.sparta.logistics.auth.libs.security;
+package com.sparta.logistics.auth.libs.jwt;
 
 import com.sparta.logistics.auth.domain.model.UserRole;
 import io.jsonwebtoken.Claims;
